[opkgc: build] Small fixes for SuSE (opensuse) packaging.
git-svn-id: http://svn.oscar.openclustergroup.org/oscar/pkgsrc@10116 18f9d06c-6ae2-0310-8438-938a554feff9

+++ build_rpm.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
-yum -y install python-devel xmlto autoconf automake
+if test -f /usr/bin/yum
+then
+    yum -y install python-devel xmlto autoconf automake
+else
+    yast -i python-devel xmlto autoconf automake
+fi
+
 ./autogen.sh
 ./configure
 make dist
